Create FantasyDraftTracker2019
This program parses over a CSV file containing fantasy football average draft position (ADP) data.  Enter the names of players as you complete your fantasy football draft, and this program will track which players you have drafted as well as provide you with the next 5 best players at each position when it is your turn to draft.  The program will also offer a suggestion for the best player available on the board, regardless of position.